# app/routers/chat.py
"""
聊天路由 - FastAPI 重构版
"""
import logging
from datetime import datetime

from fastapi import APIRouter, Depends, HTTPException, Query, Body
from sqlalchemy.orm import Session

from app.database import get_db
from app.dependencies import get_current_user
from app.models.models import User, Conversation, Message, ConversationParticipant, Friendship
from app.socket_handlers import is_user_online, get_online_users
from app.ws_manager import ws_manager

logger = logging.getLogger(__name__)
router = APIRouter()


@router.get("/conversations")
def get_conversations(
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """获取当前用户的所有会话列表"""
    from sqlalchemy import func, and_

    conversations = (
        db.query(Conversation)
        .filter(
            (Conversation.user1_id == user.id) | (Conversation.user2_id == user.id)
        )
        .order_by(func.coalesce(Conversation.last_message_at, Conversation.created_at).desc())
        .all()
    )

    if not conversations:
        return {"conversations": [], "total": 0}

    conv_ids = [c.id for c in conversations]

    max_time_subq = (
        db.query(
            Message.conversation_id,
            func.max(Message.created_at).label("max_time"),
        )
        .filter(Message.conversation_id.in_(conv_ids))
        .group_by(Message.conversation_id)
        .subquery()
    )

    last_messages = (
        db.query(Message)
        .join(
            max_time_subq,
            and_(
                Message.conversation_id == max_time_subq.c.conversation_id,
                Message.created_at == max_time_subq.c.max_time,
            ),
        )
        .all()
    )
    last_msg_map = {m.conversation_id: m for m in last_messages}

    unread_counts = (
        db.query(Message.conversation_id, func.count(Message.id).label("cnt"))
        .filter(
            Message.conversation_id.in_(conv_ids),
            Message.is_read == False,
            Message.sender_id != user.id,
        )
        .group_by(Message.conversation_id)
        .all()
    )
    unread_map = {row[0]: row[1] for row in unread_counts}

    result = []
    for conv in conversations:
        other_user = conv.user2 if conv.user1_id == user.id else conv.user1
        last_message = last_msg_map.get(conv.id)
        result.append({
            "conversation": conv.to_dict(),
            "other_user": other_user.to_dict(),
            "last_message": last_message.to_dict() if last_message else None,
            "unread_count": unread_map.get(conv.id, 0),
            "is_online": is_user_online(other_user.id),
        })

    return {"conversations": result, "total": len(result)}


@router.get("/conversations/{user_id}")
def get_conversation_with_user(
    user_id: int,
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """获取与特定用户的会话（如不存在则自动创建）"""
    other_user = db.query(User).filter(User.id == user_id).first()
    if not other_user:
        raise HTTPException(status_code=404, detail="User not found")

    uid1, uid2 = min(user.id, user_id), max(user.id, user_id)
    conversation = db.query(Conversation).filter(
        Conversation.user1_id == uid1, Conversation.user2_id == uid2
    ).first()

    if not conversation:
        # 自动创建会话 + 参与者（同一个事务）
        conversation = Conversation(user1_id=uid1, user2_id=uid2)
        db.add(conversation)
        db.flush()  # 获取 conversation.id
        participant1 = ConversationParticipant(conversation_id=conversation.id, user_id=uid1)
        participant2 = ConversationParticipant(conversation_id=conversation.id, user_id=uid2)
        db.add_all([participant1, participant2])
        db.commit()
        db.refresh(conversation)

    messages = (
        db.query(Message)
        .filter(Message.conversation_id == conversation.id)
        .order_by(Message.created_at.asc())
        .limit(50)
        .all()
    )

    return {
        "conversation": conversation.to_dict(),
        "other_user": other_user.to_dict(),
        "messages": [msg.to_dict() for msg in messages],
        "is_online": is_user_online(user_id),
    }


@router.get("/conversations/{conversation_id}/messages")
def get_messages(
    conversation_id: int,
    page: int = Query(1, ge=1),
    per_page: int = Query(50, ge=1, le=200),
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """获取会话的消息历史（分页）"""
    conversation = db.query(Conversation).filter(Conversation.id == conversation_id).first()
    if not conversation:
        raise HTTPException(status_code=404, detail="Conversation not found")
    if conversation.user1_id != user.id and conversation.user2_id != user.id:
        raise HTTPException(status_code=403, detail="Unauthorized")

    messages_query = (
        db.query(Message)
        .filter(Message.conversation_id == conversation_id)
        .order_by(Message.created_at.desc())
    )
    total = messages_query.count()
    messages = messages_query.offset((page - 1) * per_page).limit(per_page).all()
    pages = (total + per_page - 1) // per_page if total > 0 else 0

    return {
        "messages": [msg.to_dict() for msg in messages],
        "total": total,
        "pages": pages,
        "current_page": page,
        "per_page": per_page,
    }


@router.post("/conversations/{conversation_id}/messages")
async def send_message(
    conversation_id: int,
    payload: dict = Body(...),
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """HTTP 发送消息（WebSocket 降级备选）
    接收 {content, message_type, media_url, related_id}
    """
    conversation = db.query(Conversation).filter(Conversation.id == conversation_id).first()
    if not conversation:
        raise HTTPException(status_code=404, detail="Conversation not found")
    if conversation.user1_id != user.id and conversation.user2_id != user.id:
        raise HTTPException(status_code=403, detail="Unauthorized")

    content = payload.get("content", "")
    message_type = payload.get("message_type", "text")
    media_url = payload.get("media_url")
    related_id = payload.get("related_id")

    if not content and message_type == "text":
        raise HTTPException(status_code=400, detail="content is required for text messages")

    message = Message(
        conversation_id=conversation_id,
        sender_id=user.id,
        content=content,
        message_type=message_type,
        media_url=media_url,
        related_id=related_id,
    )

    try:
        db.add(message)
        # 更新会话最后消息时间
        conversation.last_message_at = datetime.utcnow()
        db.commit()
        db.refresh(message)
        logger.info(f"Message sent in conversation {conversation_id} by user {user.id}")

        # --- WebSocket 实时推送 ---
        # 推送给接收方：未读消息总数
        receiver_id = conversation.user2_id if conversation.user1_id == user.id else conversation.user1_id
        receiver_unread = (
            db.query(Message)
            .filter(
                Message.conversation_id == conversation_id,
                Message.is_read == False,
                Message.sender_id != receiver_id,
            )
            .count()
        )
        await ws_manager.send(receiver_id, {
            "type": "new_message",
            **message.to_dict(),
            "unread_count": receiver_unread,
        })
        # 也推送给发送者：同步自己的未读（通常为0）
        sender_unread = (
            db.query(Message)
            .filter(
                Message.conversation_id == conversation_id,
                Message.is_read == False,
                Message.sender_id != user.id,
            )
            .count()
        )
        await ws_manager.send(user.id, {
            "type": "new_message",
            **message.to_dict(),
            "unread_count": sender_unread,
        })

        return {"message": "Message sent", "data": message.to_dict()}
    except Exception as e:
        db.rollback()
        raise HTTPException(status_code=500, detail=str(e))
@router.post("/conversations/{conversation_id}/mark-read")
async def mark_conversation_as_read(
    conversation_id: int,
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """标记会话中的所有消息为已读"""
    conversation = db.query(Conversation).filter(Conversation.id == conversation_id).first()
    if not conversation:
        raise HTTPException(status_code=404, detail="Conversation not found")
    if conversation.user1_id != user.id and conversation.user2_id != user.id:
        raise HTTPException(status_code=403, detail="Unauthorized")

    unread = (
        db.query(Message)
        .filter(
            Message.conversation_id == conversation_id,
            Message.is_read == False,
            Message.sender_id != user.id,
        )
        .all()
    )

    for m in unread:
        m.is_read = True

    try:
        db.commit()

        # --- WebSocket 实时推送 ---
        # 计算该用户剩余的全局未读数量
        all_conv_ids = [c.id for c in db.query(Conversation).filter(
            (Conversation.user1_id == user.id) | (Conversation.user2_id == user.id)
        ).all()]
        total_unread = (
            db.query(Message)
            .filter(
                Message.conversation_id.in_(all_conv_ids),
                Message.is_read == False,
                Message.sender_id != user.id,
            )
            .count()
        )
        await ws_manager.send(user.id, {
            "type": "conversation_read",
            "conversation_id": conversation_id,
            "unread_count": total_unread,
        })

        return {"message": "All messages marked as read", "marked_count": len(unread)}
    except Exception as e:
        db.rollback()
        raise HTTPException(status_code=500, detail=str(e))


@router.get("/users/online")
def get_online_users_list(
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """获取在线用户列表"""
    online_ids = get_online_users()
    if not online_ids:
        return {"online_users": [], "total": 0}

    users = db.query(User).filter(User.id.in_(online_ids)).all()
    return {"online_users": [u.to_dict() for u in users], "total": len(users)}


@router.get("/users/{user_id}/status")
def get_user_status(
    user_id: int,
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """获取用户在线状态"""
    target = db.query(User).filter(User.id == user_id).first()
    if not target:
        raise HTTPException(status_code=404, detail="User not found")

    return {"user_id": user_id, "is_online": is_user_online(user_id), "user": target.to_dict()}


@router.get("/unread-count")
def get_unread_count(
    user: User = Depends(get_current_user),
    db: Session = Depends(get_db),
):
    """获取未读消息总数"""
    conversations = db.query(Conversation).filter(
        (Conversation.user1_id == user.id) | (Conversation.user2_id == user.id)
    ).all()

    total_unread = 0
    for conv in conversations:
        count = (
            db.query(Message)
            .filter(
                Message.conversation_id == conv.id,
                Message.is_read == False,
                Message.sender_id != user.id,
            )
            .count()
        )
        total_unread += count

    return {"total_unread": total_unread}